+++ addon/globalPlugins/voiceToggle/consts.py
@@ -9,6 +9,7 @@ addonHandler.initTranslation()
 UPDATE_API_URL = "http://api.adamsamec.cz/nvda/VoiceToggle/Update.json"
 TEMP_DIR = "..\\..\\temp\\"
 
+# Translators: Label for the silence synthesizer in the voices listbox or synthesizers and voices comboboxes
 SILENCE_VOICE_NAME = _("Silence")
 
 NORMAL_PROFILE_NAME = "[normal]"